+++ src/com/test/bookjuck/dao/BookDAO.java
@@ -1,15 +1,15 @@
 package com.test.bookjuck.dao;
 
-
+import java.beans.Statement;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.HashMap;
 
 import com.test.bookjuck.DBUtil;
 import com.test.bookjuck.dto.BookDTO;
+
 
 
 
@@ -36,7 +36,7 @@ public class BookDAO {
 		}
 	}
 	
-	
+	//검색구현
 	public ArrayList<BookDTO> list (HashMap<String, String> map){
 		
 		try {
@@ -102,6 +102,189 @@ public class BookDAO {
 		return null;
 	
 	}
+	
+	//베스트셀러 list
+	public ArrayList<BookDTO> bestSeller (HashMap<String, String> map){
+		
+		System.out.println("select");
+		
+		try {
+			
+			String where ="";
+			String sql = "";
+			
+			
+			if (map.get("yearStart") != null && map.get("yearEnd") != null
+				&& map.get("monthStart") != null && map.get("monthEnd") != null
+				&& map.get("dayStart") != null && map.get("dayEnd") != null) 
+			{
+				String start = "'" + map.get("yearStart") + map.get("monthStart") + map.get("dayStart") +"'";
+				String end = "'" + map.get("yearEnd") + map.get("monthEnd") + map.get("dayEnd") +"'";
+				
+				where = String.format("where seq between 1 and 10 and pubdate between %s and %s", start, end);
+				sql = String.format("select * from vwbestseller %s order by salerank asc", where);
+				System.out.println(sql); 
+				System.out.println(start);
+				System.out.println(end);
+				
+				
+			}
+			
+			System.out.println(where);
+			System.out.println(sql);
+					
+			
+			pstat = conn.prepareStatement(sql);
+			rs = pstat.executeQuery();
+			
+			//ResultSet -> ArrayList<DTO>
+			ArrayList<BookDTO> list = new ArrayList<BookDTO>();
+			
+			while(rs.next()) {
+				
+				BookDTO dto = new BookDTO();
+				
+				dto.setSeq(rs.getString("seq"));
+				dto.setTitle(rs.getString("title"));
+				dto.setPublisher(rs.getString("publisher"));
+				dto.setPrice(rs.getInt("price"));
+				dto.setPubDate(rs.getString("pubdate"));
+				dto.setCopy(rs.getString("copy"));
+				dto.setImage(rs.getString("image"));
+				dto.setTotalSale(rs.getString("totalsale"));
+				dto.setSaleRank(rs.getString("salerank"));
+				dto.setAuthor(rs.getString("author"));
+				System.out.println(rs.getString("title"));
+				
+				
+				list.add(dto);
+				
+				
+			}
+			
+			return list;
+					
+		} catch (Exception e) {
+			System.out.println(e);
+		}
+		
+		
+		
+		return null;
+		
+	} 
+	
+	//신간도서 list
+	public ArrayList<BookDTO> newBook (HashMap<String, String> map){
+		
+		try {
+			
+			String where ="";
+			String sql = "";
+			
+			
+			if (map.get("yearStart") != null && map.get("yearEnd") != null
+				&& map.get("monthStart") != null && map.get("monthEnd") != null
+				&& map.get("dayStart") != null && map.get("dayEnd") != null) 
+			{
+				String start = "'" + map.get("yearStart") + map.get("monthStart") + map.get("dayStart") +"'";
+				String end = "'" + map.get("yearEnd") + map.get("monthEnd") + map.get("dayEnd") +"'";
+				
+				where = String.format("where rank between 1 and 10 and pubdate between %s and %s", start, end);
+				sql = String.format("select * from vwnewbook %s", where);
+				System.out.println(sql); 
+				System.out.println(start);
+				System.out.println(end);
+							
+			}
+			
+			System.out.println(where);
+			System.out.println(sql);
+					
+			
+			pstat = conn.prepareStatement(sql);
+			rs = pstat.executeQuery();
+			
+			//ResultSet -> ArrayList<DTO>
+			ArrayList<BookDTO> list = new ArrayList<BookDTO>();
+			
+			while(rs.next()) {
+				
+				BookDTO dto = new BookDTO();
+				
+				dto.setSeq(rs.getString("seq"));
+				dto.setTitle(rs.getString("title"));
+				dto.setPublisher(rs.getString("publisher"));
+				dto.setPrice(rs.getInt("price"));
+				dto.setPubDate(rs.getString("pubdate"));
+				dto.setCopy(rs.getString("copy"));
+				dto.setImage(rs.getString("image"));
+				dto.setAuthor(rs.getString("author"));
+				dto.setRank(rs.getString("rank"));
+				System.out.println(rs.getString("title"));
+				
+				
+				list.add(dto);
+				
+				
+			}
+			
+			return list;
+					
+		} catch (Exception e) {
+			System.out.println(e);
+		}
+		
+		
+		
+		return null;
+		
+	} 
+		
+	public ArrayList<BookDTO> monthlyBestSeller() {
+		
+		
+		try {
+			
+			String sql = "select * from vwbestseller where pubdate between trunc(sysdate, 'mm') and last_day(sysdate)";
+			
+			
+					
+			
+			pstat = conn.prepareStatement(sql);
+			rs = pstat.executeQuery();
+			
+			
+			ArrayList<BookDTO> list = new ArrayList<BookDTO>();
+			
+			while (rs.next()) {
+				
+				BookDTO dto = new BookDTO();
+				
+				dto.setImage(rs.getString("image"));
+				dto.setTitle(rs.getString("title"));
+				dto.setCopy(rs.getString("copy"));
+				System.out.println(rs.getString("title"));
+							
+				
+				list.add(dto);
+				System.out.println(list);
+				System.out.println(list.size());
+				
+			}
+			
+			return list;
+			
+			
+		} catch (Exception e) {
+			System.out.println(e);
+		}
+		
+		
+		return null;
+		
+	}
+	
 	
 
 }
