--- src/com/test/bookjuck/dtosave/BookDTO.java
@@ -1,151 +1,0 @@
-package com.test.bookjuck.dtosave;
-
-public class BookDTO { //컬럼값이 유사하므로 EbookDTO도 함께 공용
-	
-	private String seq;
-	private int seqAuthor;
-	private int seqSCategory;
-	private String title; 
-	private String publisher; 
-	private int price; 
-	private int saleprice; 
-	private String pubdate; 
-	private String summary;
-	private String isbn; 
-	private String copy; 
-	private String image;
-	private String contents; 
-	private int page;
-	
-	private String Author;  
-	private String lcategory;
-	
-	private String totalSale;
-	private String saleRank;
-	
-	private String rank;
-	
-	
-	public String getSeq() {
-		return seq;
-	}
-	public void setSeq(String seq) {
-		this.seq = seq;
-	}
-	public int getSeqAuthor() {
-		return seqAuthor;
-	}
-	public void setSeqAuthor(int seqAuthor) {
-		this.seqAuthor = seqAuthor;
-	}
-	public int getSeqSCategory() {
-		return seqSCategory;
-	}
-	public void setSeqSCategory(int seqSCategory) {
-		this.seqSCategory = seqSCategory;
-	}
-	public String getTitle() {
-		return title;
-	}
-	public void setTitle(String title) {
-		this.title = title;
-	}
-	public String getPublisher() {
-		return publisher;
-	}
-	public void setPublisher(String publisher) {
-		this.publisher = publisher;
-	}
-	public int getPrice() {
-		return price;
-	}
-	public void setPrice(int price) {
-		this.price = price;
-	}
-	public int getSaleprice() {
-		return saleprice;
-	}
-	public void setSaleprice(int saleprice) {
-		this.saleprice = saleprice;
-	}
-	public String getPubdate() {
-		return pubdate;
-	}
-	public void setPubdate(String pubdate) {
-		this.pubdate = pubdate;
-	}
-	public String getSummary() {
-		return summary;
-	}
-	public void setSummary(String summary) {
-		this.summary = summary;
-	}
-	public String getIsbn() {
-		return isbn;
-	}
-	public void setIsbn(String isbn) {
-		this.isbn = isbn;
-	}
-	public String getCopy() {
-		return copy;
-	}
-	public void setCopy(String copy) {
-		this.copy = copy;
-	}
-	public String getImage() {
-		return image;
-	}
-	public void setImage(String image) {
-		this.image = image;
-	}
-	public String getContents() {
-		return contents;
-	}
-	public void setContents(String contents) {
-		this.contents = contents;
-	}
-	public int getPage() {
-		return page;
-	}
-	public void setPage(int page) {
-		this.page = page;
-	}
-	public String getAuthor() {
-		return Author;
-	}
-	public void setAuthor(String author) {
-		Author = author;
-	}
-	public String getLcategory() {
-		return lcategory;
-	}
-	public void setLcategory(String lcategory) {
-		this.lcategory = lcategory;
-	}
-	public String getTotalSale() {
-		return totalSale;
-	}
-	public void setTotalSale(String totalSale) {
-		this.totalSale = totalSale;
-	}
-	public String getSaleRank() {
-		return saleRank;
-	}
-	public void setSaleRank(String saleRank) {
-		this.saleRank = saleRank;
-	}
-	public String getRank() {
-		return rank;
-	}
-	public void setRank(String rank) {
-		this.rank = rank;
-	} 
-	
-	
-	
-	
-	
-	
-	
-
-}
