--- src/com/test/bookjuck/dtosave/EbookDTO.java
@@ -1,108 +1,0 @@
-package com.test.bookjuck.dtosave;
-
-public class EbookDTO {
-	
-	private int seq;
-	private int seqAuthor;
-	private int seqSCategory;
-	private String title; 
-	private String publisher; 
-	private int price; 
-	private int saleprice; 
-	private String pubDate; 
-	private String isbn; 
-	private String copy; 
-	private String image;
-	private String contents; 
-	private String eFile;
-	private String intro;
-	
-	public int getSeq() {
-		return seq;
-	}
-	public void setSeq(int seq) {
-		this.seq = seq;
-	}
-	public int getSeqAuthor() {
-		return seqAuthor;
-	}
-	public void setSeqAuthor(int seqAuthor) {
-		this.seqAuthor = seqAuthor;
-	}
-	public int getSeqSCategory() {
-		return seqSCategory;
-	}
-	public void setSeqSCategory(int seqSCategory) {
-		this.seqSCategory = seqSCategory;
-	}
-	public String getTitle() {
-		return title;
-	}
-	public void setTitle(String title) {
-		this.title = title;
-	}
-	public String getPublisher() {
-		return publisher;
-	}
-	public void setPublisher(String publisher) {
-		this.publisher = publisher;
-	}
-	public int getPrice() {
-		return price;
-	}
-	public void setPrice(int price) {
-		this.price = price;
-	}
-	public int getSaleprice() {
-		return saleprice;
-	}
-	public void setSaleprice(int saleprice) {
-		this.saleprice = saleprice;
-	}
-	public String getPubDate() {
-		return pubDate;
-	}
-	public void setPubDate(String pubDate) {
-		this.pubDate = pubDate;
-	}
-	public String getIsbn() {
-		return isbn;
-	}
-	public void setIsbn(String isbn) {
-		this.isbn = isbn;
-	}
-	public String getCopy() {
-		return copy;
-	}
-	public void setCopy(String copy) {
-		this.copy = copy;
-	}
-	public String getImage() {
-		return image;
-	}
-	public void setImage(String image) {
-		this.image = image;
-	}
-	public String getContents() {
-		return contents;
-	}
-	public void setContents(String contents) {
-		this.contents = contents;
-	}
-	public String geteFile() {
-		return eFile;
-	}
-	public void seteFile(String eFile) {
-		this.eFile = eFile;
-	}
-	public String getIntro() {
-		return intro;
-	}
-	public void setIntro(String intro) {
-		this.intro = intro;
-	}
-	
-	
-	
-
-}
