+++ src/com/test/bookjuck/member/book/MonthlyBestSeller.java
@@ -10,7 +10,10 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import com.test.bookjuck.dtosave.BookDTO;
+import com.test.bookjuck.dao.BookDAO;
+import com.test.bookjuck.dto.BookDTO;
+
+
 
 @WebServlet("/member/book/MonthlyBestSeller.do")
 public class MonthlyBestSeller extends HttpServlet {
@@ -29,7 +32,7 @@ public class MonthlyBestSeller extends HttpServlet {
 		for (BookDTO dto : list) {
 
 			// 날짜에서 시간 잘라내기
-			dto.setPubdate(dto.getPubdate().substring(0, 10));
+			dto.setPubDate(dto.getPubDate().substring(0, 10));
 
 			// 제목, 카피 너무 길면 자르기
 			if (dto.getTitle().length() > 30) {
