+++ src/com/test/bookjuck/search/SearchView.java
@@ -11,10 +11,10 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import com.test.bookjuck.dtosave.BookDTO;
-import com.test.bookjuck.dtosave.RealTimeSearchDAO;
-import com.test.bookjuck.dtosave.RealTimeSearchDTO;
-import com.test.bookjuck.member.book.BookDAO;
+import com.test.bookjuck.dao.BookDAO;
+import com.test.bookjuck.dao.RealTimeSearchDAO;
+import com.test.bookjuck.dto.BookDTO;
+import com.test.bookjuck.dto.RealTimeSearchDTO;
 
 
 @WebServlet("/common/searchview.do")
@@ -60,7 +60,7 @@ public class SearchView extends HttpServlet {
 			  
 		  
 		  //날짜에서 시간 잘라내기 
-		  dto.setPubdate(dto.getPubdate().substring(0,10));
+		  dto.setPubDate(dto.getPubDate().substring(0,10));
 		  
 		  //제목, 카피 너무 길면 자르기 
 		  if (dto.getTitle().length() > 30) {
